replace color-shift-hsl function to services.update themes values
